FIQ in secure world
Receives FIQ in secure world instead of in the secure monitor.

When NS bit is set in SCR the FIQ bit is set at the same time to receive
FIQ in monitor mode while in non-secure mode.

When NS bit is cleared in SCR the FIQ bit is cleared at the same time to
receive the FIQ handler of secure world.

This change simplifies the secure monitor.